+++ language_server/src/logger.cpp
@@ -12,36 +12,22 @@
  *   Broadcom, Inc. - initial API and implementation
  */
 
-
-#include <cstdlib>
-#define __STDC_WANT_LIB_EXT1__ 1
-#include <iostream>
-#include <time.h>
-
 #include "logger.h"
+
+#include <iostream>
+#include <string>
+
+#include "utils/time.h"
 
 using namespace hlasm_plugin::language_server;
 
 std::string current_time()
 {
-    std::string curr_time;
-    // Current date/time based on current time
-    time_t now = time(0);
-    // Convert current time to string
-
-#if __STDC_LIB_EXT1__ || _MSVC_LANG
-    char cstr[50];
-    ctime_s(cstr, sizeof cstr, &now);
-    curr_time.assign(cstr);
-#else
-    curr_time.assign(ctime(&now));
-#endif
-
-    // Last charactor of currentTime is "\n", so remove it
-    if (!curr_time.empty())
-        curr_time.pop_back();
-
-    return curr_time;
+    auto t = hlasm_plugin::utils::timestamp::now();
+    if (!t.has_value())
+        return "<unknown time>";
+    else
+        return t->to_string();
 }
 
 void logger::log(std::string_view data)