+++ parser_library/src/context/hlasm_context.cpp
@@ -25,6 +25,7 @@
 #include "lexing/lexer.h"
 #include "lexing/tools.h"
 #include "using.h"
+#include "utils/time.h"
 
 namespace hlasm_plugin::parser_library::context {
 
@@ -129,7 +130,7 @@ macro_data_ptr create_macro_data(std::vector<std::string> value)
 
 macro_data_ptr create_macro_data(std::unique_ptr<macro_param_data_single_dynamic> value) { return value; }
 
-template<typename SYSTEM_VARIABLE_TYPE, typename DATA>
+template<typename SYSTEM_VARIABLE_TYPE = system_variable, typename DATA>
 std::pair<id_index, sys_sym_ptr> create_system_variable(id_index id, DATA mac_data, bool is_global)
 {
     return { id, std::make_shared<SYSTEM_VARIABLE_TYPE>(id, create_macro_data(std::move(mac_data)), is_global) };
@@ -143,15 +144,13 @@ void hlasm_context::add_system_vars_to_scope(code_scope& scope)
         {
             auto sect_name = ord_ctx.current_section() ? ord_ctx.current_section()->name : id_index();
 
-            scope.system_variables.insert(
-                create_system_variable<system_variable, std::string>(id_index("SYSECT"), sect_name.to_string(), false));
+            scope.system_variables.insert(create_system_variable(id_index("SYSECT"), sect_name.to_string(), false));
         }
 
         {
             std::string value = left_pad(std::to_string(SYSNDX_), 4, '0');
 
-            scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSNDX"), std::move(value), false));
+            scope.system_variables.insert(create_system_variable(id_index("SYSNDX"), std::move(value), false));
         }
 
         {
@@ -178,8 +177,7 @@ void hlasm_context::add_system_vars_to_scope(code_scope& scope)
                 }
             }
 
-            scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSSTYP"), std::move(value), false));
+            scope.system_variables.insert(create_system_variable(id_index("SYSSTYP"), std::move(value), false));
         }
 
         {
@@ -191,14 +189,13 @@ void hlasm_context::add_system_vars_to_scope(code_scope& scope)
             }
 
             scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSLOC"), std::move(location_counter_name), false));
+                create_system_variable(id_index("SYSLOC"), std::move(location_counter_name), false));
         }
 
         {
             std::string value = std::to_string(scope_stack_.size() - 1);
 
-            scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSNEST"), std::move(value), false));
+            scope.system_variables.insert(create_system_variable(id_index("SYSNEST"), std::move(value), false));
         }
 
         {
@@ -217,13 +214,18 @@ void hlasm_context::add_system_vars_to_scope(code_scope& scope)
         }
 
         {
-            scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSIN_DSN"), asm_options_.sysin_dsn, false));
+            scope.system_variables.insert(create_system_variable(id_index("SYSIN_DSN"), asm_options_.sysin_dsn, false));
         }
 
         {
             scope.system_variables.insert(
-                create_system_variable<system_variable>(id_index("SYSIN_MEMBER"), asm_options_.sysin_member, false));
+                create_system_variable(id_index("SYSIN_MEMBER"), asm_options_.sysin_member, false));
+        }
+
+        {
+            scope.system_variables.insert(create_system_variable(id_index("SYSCLOCK"),
+                utils::timestamp::now().value_or(utils::timestamp(1900, 1, 1)).to_string(),
+                false));
         }
     }
 }
@@ -242,36 +244,35 @@ void hlasm_context::add_global_system_vars(code_scope& scope)
     if (!is_in_macro())
     {
         {
-            auto tmp_now = std::time(0);
-            auto now = std::localtime(&tmp_now);
+            auto now = utils::timestamp::now().value_or(utils::timestamp(1900, 1, 1));
 
             std::string datc_val;
             std::string date_val;
             datc_val.reserve(8);
             date_val.reserve(8);
-            auto year = std::to_string(now->tm_year + 1900);
+            auto year = std::to_string(now.year());
             datc_val.append(year);
 
-            if (now->tm_mon + 1 < 10)
+            if (now.month() < 10)
             {
                 datc_val.push_back('0');
                 date_val.push_back('0');
             }
 
-            datc_val.append(std::to_string(now->tm_mon + 1));
+            datc_val.append(std::to_string(now.month() + 1));
 
-            date_val.append(std::to_string(now->tm_mon + 1));
+            date_val.append(std::to_string(now.month() + 1));
             date_val.push_back('/');
 
-            if (now->tm_mday < 10)
+            if (now.day() < 10)
             {
                 datc_val.push_back('0');
                 date_val.push_back('0');
             }
 
-            datc_val.append(std::to_string(now->tm_mday));
+            datc_val.append(std::to_string(now.day()));
 
-            date_val.append(std::to_string(now->tm_mday));
+            date_val.append(std::to_string(now.day()));
             date_val.push_back('/');
             date_val.append(year.c_str() + 2);
 
@@ -287,13 +288,13 @@ void hlasm_context::add_global_system_vars(code_scope& scope)
 
             {
                 std::string value;
-                if (now->tm_hour < 10)
+                if (now.hour() < 10)
                     value.push_back('0');
-                value.append(std::to_string(now->tm_hour));
+                value.append(std::to_string(now.hour()));
                 value.push_back(':');
-                if (now->tm_min < 10)
+                if (now.minute() < 10)
                     value.push_back('0');
-                value.append(std::to_string(now->tm_min));
+                value.append(std::to_string(now.minute()));
 
                 globals_.insert(create_system_variable<system_variable>(id_index("SYSTIME"), std::move(value), true));
             }