+++ parser_library/test/debugging/debugger_test.cpp
@@ -260,6 +260,7 @@ struct frame_vars_ignore_sys_vars : public frame_vars
         this->locals["&SYSMAC"];
         this->locals["&SYSIN_DSN"];
         this->locals["&SYSIN_MEMBER"];
+        this->locals["&SYSCLOCK"];
     }
 };
 
@@ -516,6 +517,7 @@ TEST(debugger, test)
                 { "&SYSIN_DSN", "" },
                 { "&SYSIN_MEMBER", "" },
                 { "&VAR", "13" },
+                { "&SYSCLOCK", test_var_value() },
             },
             {} // empty ord symbols
             ));